+++ admin_ui/src/components/SidebarNav.vue
@@ -33,7 +33,6 @@
             </span>
         </p>
         <FormNav v-show="!isHiddenForms" />
-
         <p
             class="opaque"
             v-if="chartConfigs.length > 0"
@@ -50,27 +49,46 @@
             </span>
         </p>
         <ChartsNav v-show="!isHiddenCharts" />
+        <p
+            class="opaque"
+            v-if="Object.keys(customLinks).length > 0"
+            v-on:click="isHiddenLinks = !isHiddenLinks"
+        >
+            <font-awesome-icon icon="link" />{{ $t("Links") }}
+            <span style="float: right">
+                <font-awesome-icon
+                    icon="angle-down"
+                    title="Show links"
+                    v-if="isHiddenLinks"
+                />
+                <font-awesome-icon icon="angle-up" title="Hide links" v-else />
+            </span>
+        </p>
+        <LinksNav v-show="!isHiddenLinks" />
     </div>
 </template>
 
-<script>
+<script lang="ts">
 import Vue from "vue"
 import TableNav from "./TableNav.vue"
 import FormNav from "./FormNav.vue"
 import ChartsNav from "./ChartsNav.vue"
+import LinksNav from "./LinksNav.vue"
 
 export default Vue.extend({
     data() {
         return {
             isHiddenTables: false,
             isHiddenForms: false,
-            isHiddenCharts: false
+            isHiddenCharts: false,
+            isHiddenLinks: false
         }
     },
     components: {
         TableNav,
         FormNav,
-        ChartsNav
+        ChartsNav,
+        LinksNav
     },
     computed: {
         formConfigs() {
@@ -78,6 +96,9 @@ export default Vue.extend({
         },
         chartConfigs() {
             return this.$store.state.chartConfigs
+        },
+        customLinks() {
+            return this.$store.state.customLinks
         }
     }
 })