+++ admin_ui/src/store.ts
@@ -36,7 +36,8 @@ export default new Vuex.Store({
         formConfigs: [] as i.FormConfig[],
         chartConfigs: [] as i.ChartConfig[],
         user: undefined,
-        loadingStatus: false
+        loadingStatus: false,
+        customLinks: {}
     },
     mutations: {
         updateTableGroups(state, value) {
@@ -99,6 +100,9 @@ export default new Vuex.Store({
         },
         updateLoadingStatus(state, value) {
             state.loadingStatus = value
+        },
+        updateCustomLinks(state, value) {
+            state.customLinks = value
         }
     },
     actions: {
@@ -138,6 +142,10 @@ export default new Vuex.Store({
         async fetchTableGroups(context) {
             const response = await axios.get(`${BASE_URL}tables/grouped/`)
             context.commit("updateTableGroups", response.data)
+        },
+        async fetchCustomLinks(context) {
+            const response = await axios.get(`${BASE_URL}links/`)
+            context.commit("updateCustomLinks", response.data)
         },
         async fetchCount(context) {
             const tableName = context.state.currentTableName